+++ packages/remix/lib/src/components/sidebar/sidebar.dart
@@ -1,14 +1,17 @@
 library remix_sidebar;
 
+import 'dart:async';
 import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mix/mix.dart';
 import 'package:mix_annotations/mix_annotations.dart';
+import 'package:naked_ui/naked_ui.dart';
 
 import '../../utilities/remix_style.dart';
 import '../toggle/toggle.dart';
+import '../tooltip/tooltip.dart';
 
 part 'sidebar_spec.dart';
 part 'sidebar_widget.dart';